Moved network-WTF tests to its own folder.
